create base class for CondaIndexCache
