+++ tests/test_sqlitecache.py
@@ -10,6 +10,7 @@ from pathlib import Path
 
 import pytest
 
+from conda_index.index.cache import _cache_post_install_details, _cache_recipe
 from conda_index.index.common import connect
 from conda_index.index.convert_cache import (
     add_computed_name,
@@ -19,12 +20,7 @@ from conda_index.index.convert_cache import (
     merge_index_cache,
     migrate,
 )
-from conda_index.index.sqlitecache import (
-    CondaIndexCache,
-    _cache_post_install_details,
-    _cache_recipe,
-    _clear_newline_chars,
-)
+from conda_index.index.sqlitecache import CondaIndexCache, _clear_newline_chars
 from conda_index.utils import DEFAULT_SUBDIRS
